safari-534.53-branch/LayoutTests/fast/css/counters/counter-increment-000-expected.txt

This is the WebKit version of CSS 2.1 Test Suite: dynamic changes to 'counter-increment'.

Before the dynamic change:
PASS layoutTestController.counterValueForElementById('one') is '1'
PASS layoutTestController.counterValueForElementById('three') is '2'
PASS layoutTestController.counterValueForElementById('four') is '3'
After the dynamic change:
PASS layoutTestController.counterValueForElementById('one') is '1'
PASS layoutTestController.counterValueForElementById('two') is '2'
PASS layoutTestController.counterValueForElementById('three') is '3'
PASS layoutTestController.counterValueForElementById('four') is '4'

TEST COMPLETE
